AAAAA

+++ api/services/Assignment.js
@@ -1929,7 +1929,7 @@ var model = {
         Assignment.update({
           _id:n._id
         },{
-          intimatedLoss:parseInt(n.intimatedLoss)
+          intimatedLoss:parseFloat(n.intimatedLoss)
         }).lean().exec(
           function(err,data5){
             console.log("Done");
